refactor(agent): organize host conformance scenarios
